Feature/several small things (#380)
* Determine if a response is binary in ReverseProxyResponseWriter based on the content-type header as well

* Fixed Base64 decoding issue

* Fixed bug in deserializing of ConditionCheckResultModel.cs

* Implemented IsValidText helper method

* Added ASCII / text check in StringReplaceResponseWriter class

* Added method to extract content type from stub response

* Assed new tests for Base64ResponseWriter

* Added CommonResources

* Added ApplicationResources

* Added resources for ConfigKeys

* Added more resources

* Added StubResources

* Added more resources for condition checkers

* Added resources for all condition checkers

* Added more resources

* Added more resources

* Resource name refactoring

* Moved response handler names to resources

* Added infra and persistence resources

* Added resources for web shared project and made some small refactorings

* Test fixes

* Moved session-key constants

* Moved string-checking-keywords.ts

* Moved stub-examples.json

* Added basic infrastructure for translation

* Added infrastructure for making language configurable

* Added several translations

* Added translations for ScenarioForm

* Added resources for Scenarios component

* Added resources for Settings component

* Added resources for StubForm component

* Added resources for Stubs component

* Optimized imports

* Fixed bug in FileWatcherBuilder

* Added resources for ImportCurl and Modal component

* Moved languages and examples

* Moved translations for ImportHar component

* Moved translations for ImportOpenApi component

* Moved translations for UploadStubs component

* Added translations for some request components

* Added translations for all request components

* Moved translations for Stub.vue

* Moved stub helper titles and descriptions to translation files

* Re-added default stub to StubForm

* Moved more stuff to translations

* Added Markdown option to translation and translated more components

* Save fix in ScenarioForm

* Moved translations for ScenarioSelector.vue

* Added target option for markdown links

* Added entry for string checking

* Fixes in string-checking-keywords.ts

* Simplification in string-checking-keywords.ts

* Added translations for TenantSelector.vue

* Added translations for VariableHandlerSelector.vue

* Removed resources object

* Moved renderDocLink function to util class

* Moved translations for Sidebar.vue

* Moved technical constants

* Moved overlooked translations

* Moved overlooked translations

* Removed default-values.ts

* Upgraded NuGet packages

* Fixed ImageResponseWriter

* Fixed tests

---------

+++ src/HttPlaceholder.Application.Tests/StubExecution/ResponseWriters/ImageResponseWriterFacts.cs
@@ -116,8 +116,8 @@ public class ImageResponseWriterFacts
         Assert.AreEqual(1, response.Headers.Count);
         await using var ms = new MemoryStream(response.Body);
         using var image = new MagickImage(ms);
-        Assert.AreEqual(stub.Response.Image.Height, image.Height);
-        Assert.AreEqual(stub.Response.Image.Width, image.Width);
+        Assert.AreEqual((uint)stub.Response.Image.Height, image.Height);
+        Assert.AreEqual((uint)stub.Response.Image.Width, image.Width);
         Assert.AreEqual(expectedContentType, response.Headers[HeaderKeys.ContentType]);
         fileServiceMock
             .Verify(m => m.WriteAllBytesAsync(expectedCachePath, response.Body, It.IsAny<CancellationToken>()),
